Added tags to post details + minor changes
